+++ components/webui/imports/ui/SearchView/SearchView.jsx
@@ -24,7 +24,7 @@ import {unquoteString} from "/imports/utils/misc";
 
 import {LOCAL_STORAGE_KEYS} from "../constants";
 import SearchControls from "./SearchControls/SearchControls";
-import SearchResults, {VISIBLE_RESULTS_LIMIT_INITIAL} from "./SearchResults/SearchResults";
+import SearchResults, {VISIBLE_RESULTS_LIMIT_INITIAL} from "./SearchResults";
 import {computeTimelineConfig} from "./SearchResults/SearchResultsTimeline";
 import SearchStatus from "./SearchStatus";
